# chat_toeic_app/lib/core/api/chatbot_toeic.swagger.json
{"openapi":"3.0.0","info":{"title":"Chatbot TOEIC API Documentation","version":"1.0.0","description":"\nAPI documentation cho ứng dụng Chatbot TOEIC (Web + Mobile).\nSử dụng chuẩn RESTful v1.\n\n**Lưu ý cho Mobile Team:**\n- Sử dụng package `swagger_dart_code_generator` hoặc `openapi_generator` để tự động sinh code.\n- Mọi Private API đều yêu cầu Header: `Authorization: Bearer <token>`.\n      ","contact":{"name":"Development Team"}},"servers":[{"url":"http://localhost:8080","description":"Production Server"},{"url":"http://localhost:8080","description":"Development Server"}],"components":{"securitySchemes":{"BearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT","description":"JWT Access Token"}},"responses":{"SuccessResponse":{"description":"Success Response","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"success"},"message":{"type":"string","example":"Operation successful"},"data":{"type":"object","description":"Dữ liệu trả về"},"meta":{"type":"object","properties":{"timestamp":{"type":"string","format":"date-time"}}}}}}}},"UnauthorizedError":{"description":"Lỗi xác thực (401)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"},"example":{"status":"error","code":401,"message":"Unauthorized - Token is missing or invalid","errorCode":"UNAUTHORIZED"}}}},"BadRequestError":{"description":"Lỗi dữ liệu đầu vào (400)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"},"example":{"status":"error","code":400,"message":"Bad Request","details":["\"email\" is required"]}}}},"NotFoundError":{"description":"Không tìm thấy tài nguyên (404)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"},"example":{"status":"error","code":404,"message":"Resource not found"}}}}},"schemas":{"ErrorResponse":{"type":"object","properties":{"status":{"type":"string","example":"error"},"code":{"type":"integer","example":400},"message":{"type":"string","example":"Error message"},"errorCode":{"type":"string","example":"ERROR_CODE"},"details":{"type":"array","items":{"type":"string"},"example":[]},"timestamp":{"type":"string","format":"date-time"}}},"User":{"type":"object","properties":{"id":{"type":"integer","example":1},"username":{"type":"string","example":"john_doe"},"email":{"type":"string","format":"email","example":"john@example.com"},"role_id":{"type":"integer","example":1},"avatar_url":{"type":"string","nullable":true}},"required":["id","username","email"]},"Course":{"type":"object","properties":{"id":{"type":"integer","example":1},"name":{"type":"string","example":"TOEIC 750+"},"description":{"type":"string","nullable":true}},"required":["id","name"]},"Test":{"type":"object","properties":{"id":{"type":"integer","example":1},"title":{"type":"string","example":"TOEIC Reading Test 1"},"duration":{"type":"string","example":"45:00"},"participants":{"type":"integer","example":120},"commentsCount":{"type":"integer","example":5},"questionsCount":{"type":"integer","example":100},"tags":{"type":"array","items":{"type":"string"},"example":["Grammar","Reading"]}},"required":["id","title"]},"Question":{"type":"object","properties":{"id":{"type":"integer","example":1},"question":{"type":"string","example":"Select the correct option..."},"optionA":{"type":"string"},"optionB":{"type":"string"},"optionC":{"type":"string"},"optionD":{"type":"string"},"correctAnswer":{"type":"string","enum":["A","B","C","D"]},"explanation":{"type":"string","nullable":true},"typeId":{"type":"integer"},"partId":{"type":"integer","example":5},"image_url":{"type":"string","nullable":true},"audio_url":{"type":"string","nullable":true}},"required":["id","question","correctAnswer"]},"Conversation":{"type":"object","properties":{"id":{"type":"integer","example":1},"userId":{"type":"integer","example":1},"title":{"type":"string","example":"Học Part 5 TOEIC"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}},"required":["id","userId","title"]},"Message":{"type":"object","properties":{"id":{"type":"integer","example":1},"conversationId":{"type":"integer","example":1},"role":{"type":"string","enum":["user","model"]},"content":{"type":"string","example":"Giải thích giúp mình câu này..."},"createdAt":{"type":"string","format":"date-time"}},"required":["id","conversationId","role","content"]}}},"security":[{"BearerAuth":[]}],"paths":{"/api/me":{"get":{"summary":"Lấy thông tin người dùng hiện tại","description":"Trả về thông tin cơ bản của user đã đăng nhập (từ JWT token)","tags":["Auth"],"security":[{"BearerAuth":[]},{"CookieAuth":[]}],"responses":{"200":{"$ref":"#/components/responses/SuccessResponse"},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/api/logout":{"post":{"summary":"Đăng xuất","description":"Xóa token xác thực (cookies hoặc JWT)","tags":["Auth"],"responses":{"200":{"$ref":"#/components/responses/SuccessResponse"}}}},"/api/v1/auth/register":{"post":{"summary":"Đăng ký tài khoản mới","description":"Tạo một tài khoản người dùng mới với email và password","tags":["Auth (v1)"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["username","email","password"],"properties":{"username":{"type":"string","example":"johndoe"},"email":{"type":"string","format":"email","example":"john@example.com"},"password":{"type":"string","format":"password","example":"Password@123"}}}}}},"responses":{"201":{"description":"Đăng ký thành công","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"success"},"message":{"type":"string","example":"User registered successfully"},"data":{"type":"object","properties":{"id":{"type":"integer"},"username":{"type":"string"},"email":{"type":"string"}}}}}}}},"409":{"$ref":"#/components/responses/BadRequestError"},"500":{"$ref":"#/components/responses/BadRequestError"}}}},"/api/v1/auth/login":{"post":{"summary":"Đăng nhập","description":"Đăng nhập với email và password, nhận access token + refresh token","tags":["Auth (v1)"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","format":"email","example":"john@example.com"},"password":{"type":"string","format":"password","example":"Password@123"}}}}}},"responses":{"200":{"description":"Đăng nhập thành công","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"success"},"message":{"type":"string","example":"Login successful"},"data":{"type":"object","properties":{"accessToken":{"type":"string","description":"JWT access token"},"refreshToken":{"type":"string","description":"JWT refresh token"},"user":{"type":"object","properties":{"id":{"type":"integer"},"username":{"type":"string"},"email":{"type":"string"},"role_id":{"type":"integer"}}}}}}}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"500":{"$ref":"#/components/responses/BadRequestError"}}}},"/api/v1/auth/refresh":{"post":{"summary":"Làm mới access token","description":"Sử dụng refresh token để lấy access token mới","tags":["Auth (v1)"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["refreshToken"],"properties":{"refreshToken":{"type":"string","description":"Refresh token nhận được từ login","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}}}}}},"responses":{"200":{"description":"Token làm mới thành công","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"success"},"message":{"type":"string","example":"Token refreshed successfully"},"data":{"type":"object","properties":{"accessToken":{"type":"string"}}}}}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/api/v1/auth/logout":{"post":{"summary":"Đăng xuất","description":"Đăng xuất và xóa refresh token","tags":["Auth (v1)"],"requestBody":{"required":false,"content":{"application/json":{"schema":{"type":"object","properties":{"refreshToken":{"type":"string","description":"Refresh token (optional)"}}}}}},"responses":{"200":{"description":"Đăng xuất thành công"}}}},"/api/v1/auth/me":{"get":{"summary":"Lấy thông tin người dùng hiện tại","description":"Trả về thông tin cơ bản của user đã đăng nhập","tags":["Auth (v1)"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Lấy thông tin thành công","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"success"},"message":{"type":"string"},"data":{"type":"object","properties":{"id":{"type":"integer"},"username":{"type":"string"},"email":{"type":"string"},"avatar":{"type":"string"},"role_id":{"type":"integer"}}}}}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"}}}},"/api/v1/users/me/conversations":{"get":{"summary":"Lấy danh sách hội thoại của tôi","tags":["Chatbot (v1)"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Thành công"}}}},"/api/v1/conversations":{"post":{"summary":"Tạo hội thoại mới","tags":["Chatbot (v1)"],"security":[{"BearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string"}}}}}},"responses":{"201":{"description":"Thành công"}}}},"/api/v1/conversations/{id}":{"get":{"summary":"Lấy chi tiết hội thoại","tags":["Chatbot (v1)"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Thành công"}}}},"/api/v1/conversations/{conversationId}/messages":{"get":{"summary":"Lấy danh sách tin nhắn trong hội thoại","tags":["Chatbot (v1)"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"conversationId","required":true,"schema":{"type":"integer"}},{"in":"query","name":"format","schema":{"type":"string","enum":["gemini"]}}],"responses":{"200":{"description":"Thành công"}}},"post":{"summary":"Thêm tin nhắn vào hội thoại","tags":["Chatbot (v1)"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"conversationId","required":true,"schema":{"type":"integer"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["role","content"],"properties":{"role":{"type":"string","enum":["user","model"]},"content":{"type":"string"}}}}}},"responses":{"201":{"description":"Thành công"}}}},"/api/v1/conversations/{conversationId}/ask":{"post":{"summary":"Hỏi chatbot AI","tags":["Chatbot (v1)"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"conversationId","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["rawText"],"properties":{"rawText":{"type":"string"}}}}}},"responses":{"200":{"description":"Thành công"}}}},"/api/v1/courses":{"get":{"summary":"Lấy danh sách khóa học","tags":["Course (v1)"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"query","name":"include","schema":{"type":"string","enum":["tests"]},"description":"Bao gồm danh sách bài thi trong mỗi khóa học"}],"responses":{"200":{"description":"Thành công","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/responses/SuccessResponse"},{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Course"}}}}]}}}}}},"post":{"summary":"Tạo khóa học mới","tags":["Course (v1)"],"security":[{"BearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["name"],"properties":{"name":{"type":"string","example":"TOEIC 990 VIP"}}}}}},"responses":{"201":{"description":"Đã tạo","content":{"application/json":{"schema":{"$ref":"#/components/responses/SuccessResponse"}}}}}}},"/api/v1/courses/{id}":{"patch":{"summary":"Cập nhật thông tin khóa học","tags":["Course (v1)"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"}}}}}},"responses":{"200":{"description":"Thành công"}}},"delete":{"summary":"Xóa khóa học","tags":["Course (v1)"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Thành công"}}}},"/api/v1/statistics/user-tests":{"get":{"summary":"Thống kê tổng quan bài thi của user","tags":["Statistics (v1)"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Thành công"}}}},"/api/v1/statistics/parts":{"get":{"summary":"Thống kê theo từng Part TOEIC","tags":["Statistics (v1)"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Thành công"}}}},"/api/v1/statistics/accuracy-over-time":{"get":{"summary":"Thống kê độ chính xác theo thời gian","tags":["Statistics (v1)"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"query","name":"days","schema":{"type":"integer","default":30}}],"responses":{"200":{"description":"Thành công"}}}},"/api/v1/statistics/user-test-history":{"get":{"summary":"Lịch sử làm bài thi chi tiết","tags":["Statistics (v1)"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Thành công"}}}},"/api/v1/tests":{"get":{"summary":"Lấy danh sách bài thi","tags":["Test (v1)"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Thành công","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/responses/SuccessResponse"},{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Test"}}}}]}}}}}}},"/api/v1/tests/{testId}/questions":{"get":{"summary":"Lấy danh sách câu hỏi của bài thi","tags":["Test (v1)"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"testId","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Thành công"}}},"post":{"summary":"Thêm câu hỏi vào bài thi","tags":["Test (v1)"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"testId","required":true,"schema":{"type":"integer"}}],"responses":{"201":{"description":"Thành công"}}}},"/api/v1/tests/{testId}/attempts":{"post":{"summary":"Bắt đầu làm bài thi","tags":["Test (v1)"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"testId","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Thành công"}}}},"/api/v1/tests/{testId}/attempts/{attemptId}/submit":{"post":{"summary":"Nộp bài thi","tags":["Test (v1)"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"testId","required":true,"schema":{"type":"integer"}},{"in":"path","name":"attemptId","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Thành công"}}}},"/api/v1/tests/{testId}/attempts/latest/check":{"get":{"summary":"Kiểm tra trạng thái làm bài gần nhất","tags":["Test (v1)"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"testId","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Thành công"}}}},"/api/v1/tests/{testId}/attempts/history":{"get":{"summary":"Lấy lịch sử làm bài của một test","tags":["Test (v1)"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"testId","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Thành công"}}}},"/api/v1/test-attempts/{attemptId}/result":{"get":{"summary":"Lấy kết quả chi tiết của lần làm bài","tags":["Test (v1)"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"attemptId","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Thành công"}}}},"/api/v1/practice-attempts/submit":{"post":{"summary":"Nộp bài luyện tập","tags":["Test (v1)"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Thành công"}}}},"/api/v1/questions/{id}":{"patch":{"summary":"Cập nhật câu hỏi","tags":["Test (v1)"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Thành công"}}}},"/api/v1/uploads/images":{"post":{"summary":"Upload ảnh lên Cloudinary","tags":["Upload (v1)"],"security":[{"BearerAuth":[]}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"Thành công"}}}},"/api/v1/uploads/audio":{"post":{"summary":"Upload audio lên Cloudinary","tags":["Upload (v1)"],"security":[{"BearerAuth":[]}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"Thành công"}}}},"/api/v1/uploads/{publicId}":{"delete":{"summary":"Xóa file từ Cloudinary","tags":["Upload (v1)"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"publicId","required":true,"schema":{"type":"string"}},{"in":"query","name":"resourceType","schema":{"type":"string","enum":["image","video"]},"description":"image cho ảnh, video cho audio"}],"responses":{"200":{"description":"Thành công"}}}},"/api/v1/uploads/batch":{"post":{"summary":"Batch upload từ local paths (Admin use)","tags":["Upload (v1)"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Thành công"}}}},"/api/v1/uploads/validate-paths":{"post":{"summary":"Kiểm tra tính hợp lệ của local paths","tags":["Upload (v1)"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Thành công"}}}}},"tags":[{"name":"Auth (v1)","description":"Authentication endpoints (v1 API)"}]}